Conflicts:
	bower.json
	tests/webtest_tests.py
	website/templates/dashboard.mako
	website/templates/project/new.mako
	website/templates/project/project_base.mako

+++ website/addons/figshare/static/figshare-node-cfg.js
@@ -17,13 +17,14 @@
         });
 
         $('#figshareDelKey').on('click', function() {
-            bootbox.confirm(
-                'Are you sure you want to delete your Figshare access key? This will ' +
+            bootbox.confirm({
+                title: 'Remove access key?',
+                message: 'Are you sure you want to remove your Figshare access key? This will ' +
                 'revoke the ability to modify and upload files to Figshare. If ' +
                 'the associated repo is private, this will also disable viewing ' +
                 'and downloading files from Figshare.',
-                function(result) {
-                    if (result) {
+                callback: function(result) {
+                    if(result) {
                         $.ajax({
                             url: nodeApiUrl + 'figshare/oauth/',
                             type: 'DELETE',
@@ -35,7 +36,7 @@
                         });
                     }
                 }
-            );
+            });
         });
 
         $('#figshareSelectProject').on('change', function() {
@@ -65,27 +66,20 @@
 
         bootbox.prompt('Name your new file set', function(filesetName) {
             if (filesetName && filesetName.trim() != '') {
-                $.ajax({
-                    type: 'POST',
-                    url: nodeApiUrl + 'figshare/new/fileset/',
-                    contentType: 'application/json',
-                    dataType: 'json',
-                    data: JSON.stringify({
-                        name: filesetName
-                    }),
-                    success: function(response) {
-                        response.article_id = 'fileset_' + response.items[0].article_id;
-                        $select.append('<option value="' + response.article_id + '">' + filesetName + ':' + response.items[0].article_id + '</option>');
-                        $select.val(response.article_id);
-                        $('#figshareId').val(response.article_id)
-                        $('#figshareTitle').val(filesetName)
-                    },
-                    error: function() {
-                        $('#addonSettingsFigshare').find('.addon-settings-message')
-                            .text('Could not create file set')
-                            .removeClass('text-success').addClass('text-danger')
-                            .fadeOut(100).fadeIn();
-                    }
+                $.osf.postJSON(
+                    nodeApiUrl + 'figshare/new/fileset/',
+                    {name: filesetName}
+                ).done(function(response) {
+                    response.article_id = 'fileset_' + response.items[0].article_id;
+                    $select.append('<option value="' + response.article_id + '">' + filesetName + ':' + response.items[0].article_id + '</option>');
+                    $select.val(response.article_id);
+                    $('#figshareId').val(response.article_id)
+                    $('#figshareTitle').val(filesetName)
+                }).fail(function() {
+                    $('#addonSettingsFigshare').find('.addon-settings-message')
+                        .text('Could not create file set')
+                        .removeClass('text-success').addClass('text-danger')
+                        .fadeOut(100).fadeIn();
                 });
             }
         });
